DjangoDemo/typeidea/requirements.txt
-i https://pypi.douban.com/simple
django
mistune
coreapi
django-rest-framework
django-debug-toolbar
djdt_flamegraph
pympler
django-debug-toolbar-line-profiler
